refactor: first gas optimization of veOLA

+++ contracts/interfaces/IStructs.sol
@@ -71,16 +71,18 @@ interface IStructs {
     }
 
     // Structure for voting escrow points
+    // The struct size is two storage slots of 2 * uint256 (128 + 128 + 64 + 64 + 128)
     struct PointVoting {
         // w(i) = at + b (bias)
         int128 bias;
         // dw / dt = a (slope)
         int128 slope;
-        // Timestamp
-        uint256 ts;
-        // Block number
-        uint256 blockNumber;
-        // Supply or account balance
-        uint256 balance;
+        // Timestamp. It will never practically be bigger than 2^64 - 1
+        uint64 ts;
+        // Block number. It will not be bigger than the timestamp
+        uint64 blockNumber;
+        // Token amount. It will never practically be bigger. Initial OLA cap is 1 bn tokens, or 1e27.
+        // After 10 years, the inflation rate is 2% per year. It would take 1340+ years to reach 2^128 - 1
+        uint128 balance;
     }
 }
